+++ AterrizarSA_Grupo5/NuevoPasajeroHotelModel.cs
@@ -36,7 +36,7 @@ namespace AterrizarSA_Grupo5
         public int CantidadCamasAdultos { get; set; }
         public int CantidadCamasMenores { get; set; }
         public int CantidadCamasInfantes { get; set; }
-        public double PrecioNoche { get; set; }
+        public decimal PrecioNoche { get; set; }
         public DateTime FechaDesde { get; set; }
         public DateTime FechaHasta { get; set; }
